Struct aquascope::analysis::ir_mapper::IRMapper

pub struct IRMapper<'a, 'tcx> { /* private fields */ }

Implementations

impl<'a, 'tcx> IRMapper<'a, 'tcx>
where 'tcx: 'a,

pub fn new( tcx: TyCtxt<'tcx>, body: &'a Body<'tcx>, gather_mode: GatherMode ) -> Self

pub fn ldominates(&self, dom: Location, node: Location) -> bool

pub fn lpost_dominates(&self, pdom: Location, node: Location) -> bool

pub fn dominates(&self, dom: BasicBlock, node: BasicBlock) -> bool

pub fn post_dominates(&self, pdom: BasicBlock, node: BasicBlock) -> bool

pub fn is_terminator_switchint(&self, location: Location) -> bool

Returns true if the terminator in the location’s block is a switchInt.

pub fn let_assigned_place(&self, local: &LetStmt<'_>) -> Vec<Place<'tcx>>

pub fn get_mir_locations( &self, hir_id: HirId, depth: GatherDepth ) -> Option<MirOrderedLocations>

Produces a MirOrderedLocations which is defined as follows. The entry_block represents the BasicBlock which post-dominates all blocks in the given set of locations and conversely the exit_block dominates all blocks in the set.

This works under the assumption that there exists a global maximum in the (post-)dominator lattice.

See: https://en.wikipedia.org/wiki/Dominator_(graph_theory)