Struct aquascope::analysis::permissions::Permissions

pub struct Permissions {
    pub read: bool,
    pub write: bool,
    pub drop: bool,
}

Read, Write, and Own permissions for a single Place.

NOTE: previously, the term drop was used instead of own and this terminology remains within the source and internal documentation.

Fields
read: boolwrite: booldrop: bool
Implementations

impl Permissions

pub fn none() -> Self

pub fn all() -> Self

impl Permissions

pub fn bottom() -> Self