Struct aquascope::analysis::permissions::PermissionsData

pub struct PermissionsData {
    pub type_droppable: bool,
    pub type_writeable: bool,
    pub type_copyable: bool,
    pub is_live: bool,
    pub path_uninitialized: bool,
    pub path_moved: Option<MoveKey>,
    pub loan_read_refined: Option<LoanKey>,
    pub loan_write_refined: Option<LoanKey>,
    pub loan_drop_refined: Option<LoanKey>,
    pub permissions: Permissions,
}

Permissions and first-order provenance for permission refinement.

In contrast to Permissions, the PermissionsData stores all relevant information about what factors into the permissions. Things like declared type information, loan refinements, move refinements, etc. PermissionsData corresponds to a single Place.

Fields
type_droppable: bool

Was the type declared as droppable (i.e. an owned value)?

type_writeable: bool

Was the type declared as writeable (i.e. is it mut)?

type_copyable: bool

Is the type copyable (i.e. does it implement the Copy trait)?

is_live: bool

Is the Place currently live?

path_uninitialized: bool

Is this place uninitialized?

path_moved: Option<MoveKey>

Is the Place currently uninitialized due to a move?

loan_read_refined: Option<LoanKey>

Is a live loan removing read permissions?

loan_write_refined: Option<LoanKey>

Is a live loan removing write permissions?

loan_drop_refined: Option<LoanKey>

Is a live loan removing drop permissions?

permissions: Permissions

Computed permissions given the above information.