Module aquascope::analysis::stepper

Analysis for the “Missing-at” relations.

Structs

MirSegment

Represents a segment of the MIR control-flow graph.

PermissionsDataDiff

PermissionsDiff

PermissionsLineDisplay

A collection of PermissionsStepTable which are to be shown at the same location.

PermissionsStepTable

Table representation of the permissions difference between two locations.

Enums

PermIncludeMode

ValueStep

Statics

INCLUDE_MODE

Traits

Difference

Stepable

Functions

compute_permission_steps